Limit the number of waiting gossiper tasks
It can happen that the tasks arrive faster than the node is able
to process then.
In this case the node consumes all the available memory and crashes.
Instead of crashing, drop the gossip state applies for a while.
The state should synchronize once the node is not overloaded with tasks
again.

See https://github.com/scylladb/scylladb/issues/10967